2.0.0.2: - FFMPEGProducer optimization. - Layer pause. - FFMPEGProducer Frame seeking (timecode). - GPU scaling was not enabled, it is now. - Fixed a composite frame translation bug. - GPU based interlacing git-svn-id: https://casparcg.svn.sourceforge.net/svnroot/casparcg/server/branches/2.0.0.2@185 362d55ac-95cf-4e76-9f9a-cbaa9c17b72d

+++ core/frame/gpu_frame.h
@@ -29,12 +29,12 @@ public:
 	virtual const std::vector<short>& audio_data() const;	
 	virtual std::vector<short>& audio_data();
 
-	virtual float alpha() const;
+	virtual double alpha() const;
 	virtual void alpha(float value);
 
-	virtual float x() const;
-	virtual float y() const;
-	virtual void translate(float x, float y);
+	virtual double x() const;
+	virtual double y() const;
+	virtual void translate(double x, double y);
 
 	virtual void mode(video_mode mode);
 	virtual video_mode mode() const;
